Spring cleaning: file splitting
