+++ gif_test.go
